[Impeller] Refactor and document Android rendering related TUs.
No change in functionality. The updates:

* HeaderDoc comments have been added to the major rendering related TUs.
  However, not all comments are complete. I only documented the stuff I
  understood. For instance, it isn't clear how/why some subclasses use
  Attach/Detach while other use ProcessFrame. I intend to document as I
  go.
* Each package specific subclass is now in its own TU.
* Some subclasses used to be Skia specific. We added an Impeller variant
  but didn't rename to Skia TU. This has been patched.
