+++ shell/platform/android/BUILD.gn
@@ -84,8 +84,8 @@ source_set("flutter_shell_native_src") {
     "android_context_gl_impeller.h",
     "android_context_gl_skia.cc",
     "android_context_gl_skia.h",
-    "android_context_vulkan_impeller.cc",
-    "android_context_vulkan_impeller.h",
+    "android_context_vk_impeller.cc",
+    "android_context_vk_impeller.h",
     "android_display.cc",
     "android_display.h",
     "android_egl_surface.cc",
@@ -100,8 +100,8 @@ source_set("flutter_shell_native_src") {
     "android_surface_gl_skia.h",
     "android_surface_software.cc",
     "android_surface_software.h",
-    "android_surface_vulkan_impeller.cc",
-    "android_surface_vulkan_impeller.h",
+    "android_surface_vk_impeller.cc",
+    "android_surface_vk_impeller.h",
     "apk_asset_provider.cc",
     "apk_asset_provider.h",
     "flutter_main.cc",
@@ -110,8 +110,12 @@ source_set("flutter_shell_native_src") {
     "image_external_texture.h",
     "image_external_texture_gl.cc",
     "image_external_texture_gl.h",
-    "image_external_texture_vk.cc",
-    "image_external_texture_vk.h",
+    "image_external_texture_gl_impeller.cc",
+    "image_external_texture_gl_impeller.h",
+    "image_external_texture_gl_skia.cc",
+    "image_external_texture_gl_skia.h",
+    "image_external_texture_vk_impeller.cc",
+    "image_external_texture_vk_impeller.h",
     "image_lru.cc",
     "image_lru.h",
     "library_loader.cc",
@@ -125,8 +129,10 @@ source_set("flutter_shell_native_src") {
     "platform_view_android_jni_impl.h",
     "surface_texture_external_texture.cc",
     "surface_texture_external_texture.h",
-    "surface_texture_external_texture_gl.cc",
-    "surface_texture_external_texture_gl.h",
+    "surface_texture_external_texture_gl_impeller.cc",
+    "surface_texture_external_texture_gl_impeller.h",
+    "surface_texture_external_texture_gl_skia.cc",
+    "surface_texture_external_texture_gl_skia.h",
     "vsync_waiter_android.cc",
     "vsync_waiter_android.h",
   ]
