+++ shell/platform/android/platform_view_android.cc
@@ -14,16 +14,18 @@
 #include "flutter/shell/gpu/gpu_surface_gl_delegate.h"
 #include "flutter/shell/platform/android/android_context_gl_impeller.h"
 #include "flutter/shell/platform/android/android_context_gl_skia.h"
-#include "flutter/shell/platform/android/android_context_vulkan_impeller.h"
+#include "flutter/shell/platform/android/android_context_vk_impeller.h"
 #include "flutter/shell/platform/android/android_surface_gl_impeller.h"
 #include "flutter/shell/platform/android/android_surface_gl_skia.h"
 #include "flutter/shell/platform/android/android_surface_software.h"
-#include "flutter/shell/platform/android/image_external_texture_gl.h"
-#include "flutter/shell/platform/android/surface_texture_external_texture_gl.h"
+#include "flutter/shell/platform/android/image_external_texture_gl_impeller.h"
+#include "flutter/shell/platform/android/image_external_texture_gl_skia.h"
+#include "flutter/shell/platform/android/surface_texture_external_texture_gl_impeller.h"
+#include "flutter/shell/platform/android/surface_texture_external_texture_gl_skia.h"
 #include "fml/logging.h"
 #if IMPELLER_ENABLE_VULKAN  // b/258506856 for why this is behind an if
-#include "flutter/shell/platform/android/android_surface_vulkan_impeller.h"
-#include "flutter/shell/platform/android/image_external_texture_vk.h"
+#include "flutter/shell/platform/android/android_surface_vk_impeller.h"
+#include "flutter/shell/platform/android/image_external_texture_vk_impeller.h"
 #endif
 #include "flutter/shell/platform/android/context/android_context.h"
 #include "flutter/shell/platform/android/external_view_embedder/external_view_embedder.h"
@@ -53,9 +55,8 @@ std::unique_ptr<AndroidSurface> AndroidSurfaceFactoryImpl::CreateSurface() {
       return std::make_unique<AndroidSurfaceGLSkia>(
           std::static_pointer_cast<AndroidContextGLSkia>(android_context_));
     case AndroidRenderingAPI::kImpellerVulkan:
-      return std::make_unique<AndroidSurfaceVulkanImpeller>(
-          std::static_pointer_cast<AndroidContextVulkanImpeller>(
-              android_context_));
+      return std::make_unique<AndroidSurfaceVKImpeller>(
+          std::static_pointer_cast<AndroidContextVKImpeller>(android_context_));
   }
   FML_UNREACHABLE();
 }
@@ -75,7 +76,7 @@ static std::shared_ptr<flutter::AndroidContext> CreateAndroidContext(
           std::make_unique<impeller::egl::Display>(),
           enable_opengl_gpu_tracing);
     case AndroidRenderingAPI::kImpellerVulkan:
-      return std::make_unique<AndroidContextVulkanImpeller>(
+      return std::make_unique<AndroidContextVKImpeller>(
           enable_vulkan_validation, enable_vulkan_gpu_tracing);
     case AndroidRenderingAPI::kSkiaOpenGLES:
       return std::make_unique<AndroidContextGLSkia>(
@@ -276,14 +277,14 @@ void PlatformViewAndroid::RegisterExternalTexture(
   switch (android_context_->RenderingApi()) {
     case AndroidRenderingAPI::kImpellerOpenGLES:
       // Impeller GLES.
-      RegisterTexture(std::make_shared<SurfaceTextureExternalTextureImpellerGL>(
+      RegisterTexture(std::make_shared<SurfaceTextureExternalTextureGLImpeller>(
           std::static_pointer_cast<impeller::ContextGLES>(
               android_context_->GetImpellerContext()),
           texture_id, surface_texture, jni_facade_));
       break;
     case AndroidRenderingAPI::kSkiaOpenGLES:
       // Legacy GL.
-      RegisterTexture(std::make_shared<SurfaceTextureExternalTextureGL>(
+      RegisterTexture(std::make_shared<SurfaceTextureExternalTextureGLSkia>(
           texture_id, surface_texture, jni_facade_));
       break;
     case AndroidRenderingAPI::kSoftware:
@@ -316,7 +317,7 @@ void PlatformViewAndroid::RegisterImageTexture(
           texture_id, image_texture_entry, jni_facade_));
       break;
     case AndroidRenderingAPI::kImpellerVulkan:
-      RegisterTexture(std::make_shared<ImageExternalTextureVK>(
+      RegisterTexture(std::make_shared<ImageExternalTextureVKImpeller>(
           std::static_pointer_cast<impeller::ContextVK>(
               android_context_->GetImpellerContext()),
           texture_id, image_texture_entry, jni_facade_));
